Renaming components co-authored by Jamie Li
